requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
pydantic[email]>=2.7.0
pyjwt>=2.8.0
bcrypt>=4.1.0
httpx>=0.27.0
python-dotenv>=1.0.0
langgraph>=0.0.55
apscheduler>=3.10.0